+++ packages/analytics/analytics-utilities/src/types/explore/ai.ts
@@ -4,10 +4,13 @@ import type { BasicExploreEmptyFilterV2, BasicExploreInFilterV2, BasicExploreQue
 export const queryableAiExploreDimensions = [
   'control_plane',
   'control_plane_group',
+  'ai_gateway',
   'gateway_service',
   'consumer',
+  'ai_gateway_consumer',
+  'ai_gateway_consumer_group',
   'application',
-  'authenticated_entity',
+  'oidc_credential',
   'route',
   'ai_provider',
   'ai_response_model',
@@ -22,6 +25,10 @@ export const queryableAiExploreDimensions = [
   'ai_plugin',
   'principal',
   'cache_status',
+  'ai_gateway_mcp_server',
+  'ai_gateway_model',
+  'ai_gateway_agent',
+  'ai_gateway_data_plane_node',
 ] as const
 
 export type QueryableAiExploreDimensions = typeof queryableAiExploreDimensions[number]
@@ -42,6 +49,7 @@ export const aiExploreAggregations = [
   'total_tokens',
   'prompt_tokens',
   'completion_tokens',
+  'cached_tokens',
   'ai_request_count',
   'error_rate',
   'cost',
@@ -50,6 +58,7 @@ export const aiExploreAggregations = [
   'llm_latency_average',
   'llm_embeddings_tokens',
   'llm_embeddings_cost',
+  'time_to_first_token_average',
 ] as const
 
 export type AiExploreAggregations = typeof aiExploreAggregations[number]