+++ packages/entities/entities-plugins/src/components/free-form/shared/CredentialSecretField.vue
@@ -35,7 +35,7 @@ import { SensitiveInput } from '@kong-ui-public/entities-shared'
 import type { SensitiveInputLabels } from '@kong-ui-public/entities-shared'
 import { useField, useFieldAttrs } from './composables'
 import { generateCredentialSecret } from './utils'
-import type { BaseFieldProps } from './types'
+import type { BaseFieldProps, EmptyValue } from './types'
 
 defineOptions({
   inheritAttrs: false,
@@ -53,14 +53,14 @@ const { name, generator, ...props } = defineProps<CredentialSecretFieldProps & {
   generator?: () => string | Promise<string>
 }>()
 const emit = defineEmits<{
-  'update:modelValue': [value: string | null]
+  'update:modelValue': [value: string | EmptyValue]
 }>()
 
-const { value: fieldValue, hide, ...field } = useField<string | null>(toRef(() => name))
+const { value: fieldValue, hide, ...field } = useField<string | EmptyValue>(toRef(() => name))
 const fieldAttrs = useFieldAttrs(field.path!, toRef({ ...props, ...attrs }))
 
 function handleUpdate(value: string) {
-  fieldValue!.value = value === '' ? null : value
+  fieldValue!.value = value === '' ? field.emptyValue!.value : value
   emit('update:modelValue', fieldValue!.value)
 }
 </script>