+++ packages/analytics/analytics-config-store/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.23](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-config-store%401.4.22...%40kong-ui-public%2Fanalytics-config-store%401.4.23) (2026-08-22)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [1.4.22](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-config-store%401.4.21...%40kong-ui-public%2Fanalytics-config-store%401.4.22) (2026-08-22)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [1.4.21](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fanalytics-config-store%401.4.20...%40kong-ui-public%2Fanalytics-config-store%401.4.21) (2026-08-21)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [1.4.20](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.4.19...@kong-ui-public/analytics-config-store@1.4.20) (2026-08-20)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [1.4.19](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.4.18...@kong-ui-public/analytics-config-store@1.4.19) (2026-08-18)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [1.4.18](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.4.17...@kong-ui-public/analytics-config-store@1.4.18) (2026-08-18)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
 ## [1.4.17](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.4.16...@kong-ui-public/analytics-config-store@1.4.17) (2026-08-17)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-config-store