+++ packages/entities/entities-shared/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.65.0](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-shared%403.64.0...%40kong-ui-public%2Fentities-shared%403.65.0) (2026-08-24)
+
+### Features
+
+* **entities-plugins:** add entity name for plugin config ([#3685](https://github.com/Kong/public-ui-components/issues/3685)) ([0c461a4](https://github.com/Kong/public-ui-components/commit/0c461a4397e92b5b8abb68c9ae3ce82fd7ad003c))
+
+
+# [3.64.0](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-shared%403.63.4...%40kong-ui-public%2Fentities-shared%403.64.0) (2026-08-24)
+
+### Features
+
+* **entities-plugins:** rename governance to entitlement enforcement ([#3703](https://github.com/Kong/public-ui-components/issues/3703)) ([b7ab8d4](https://github.com/Kong/public-ui-components/commit/b7ab8d46dfd074e82311404c8d9b9c520785e80e))
+
+
+## [3.63.4](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-shared%403.63.3...%40kong-ui-public%2Fentities-shared%403.63.4) (2026-08-22)
+
+**Note:** Version bump only for package @kong-ui-public/entities-shared
+
+
+
+
+
+## [3.63.3](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-shared%403.63.2...%40kong-ui-public%2Fentities-shared%403.63.3) (2026-08-22)
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies with stable versions ([#3731](https://github.com/Kong/public-ui-components/issues/3731)) ([79e2e2d](https://github.com/Kong/public-ui-components/commit/79e2e2d53fdd6362c899464ee44a27bdf5370a14))
+
+
+## [3.63.2](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-shared%403.63.1...%40kong-ui-public%2Fentities-shared%403.63.2) (2026-08-22)
+
+**Note:** Version bump only for package @kong-ui-public/entities-shared
+
+
+
+
+
+## [3.63.1](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fentities-shared%403.63.0...%40kong-ui-public%2Fentities-shared%403.63.1) (2026-08-22)
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#3723](https://github.com/Kong/public-ui-components/issues/3723)) ([df7e955](https://github.com/Kong/public-ui-components/commit/df7e95536ba2f2ea8d46fa1de541b163b5b1c126))
+
+
+# [3.63.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.62.1...@kong-ui-public/entities-shared@3.63.0) (2026-08-21)
+
+
+### Features
+
+* **entities-gateway-services:** don't block delete on the force-delete checkbox ([#3720](https://github.com/Kong/public-ui-components/issues/3720)) ([a8f6fbc](https://github.com/Kong/public-ui-components/commit/a8f6fbc55a7a347d893c36afac4cb8d663e2e8d5))
+
+
+
+
+
 ## [3.62.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-shared@3.62.0...@kong-ui-public/entities-shared@3.62.1) (2026-08-14)